+++ ryoku/shell/quickshell/shell/modules/bar/barstyles/qsbar/variants/V2/panels/ArchUpdaterPanel.qml
@@ -1111,7 +1111,7 @@ PanelWindow {
                         onClicked: {
                             var command = [archPanel.packageUpdateCommand]
                             if (archPanel.usesRyokuUpdater) {
-                                if (archPanel.packageUpdateMode === "cli") command.push("update")
+                                command.push("update")
                             } else {
                                 command.push(
                                     "--run",